Added Font and update
Added the font I'm using for the Error message.  The machine I'm on has a built-
in mic I can't disable, so I can't test the error message right now.  I suppose
tomorrow I will comment a bit more on what I've done, but this was mostly just a
test, so It's not a big deal yet.